+++ alquiler/src/main/java/com/ceiba/alquiler/dominio/repositorio/RepositorioCliente.java
@@ -27,7 +27,7 @@ public interface RepositorioCliente {
 	 * Metodo para eliminar cliente por medio de su cedula
 	 * @param cedula, cedula a consultar para eliminar el cliente
 	 */
-	void eliminar(int cedula);
+	void eliminar(Long cedula);
 	
 	/**
 	 * Metodo que retorna la lista de cliente
